latest changes

+++ src/main/scala/beam/agentsim/agents/choice/mode/ModeChoiceLCCM.scala
@@ -242,7 +242,7 @@ class ModeChoiceLCCM(
           walkTime,
           waitTime,
           bikeTime,
-          totalCost.toDouble,
+          totalCost,
           altAndIdx._2
         )
       }
